Add Day03 problem solution
This commit adds implementation to solve the Day03 coding problem. The solution includes a csproj file, input text files and the main Program.cs logic. The solution primarily involves solving the problem in two distinct ways by finding and categorizing numbers within a list of coordinates and checking the presence of specific symbols around such numbers.

TBH this could be the ugliest solution for a AoC problem I have ever written but need to play catchup